Relocate task-contract / feedback helpers (batch 3) to actor_loop_flow
Move 10 helpers (~176 LOC of bodies, 202 net file LOC) out of turn.rs
into actor_loop_flow.rs where their production callers already live.
Pure code motion + orphaned doc-comment cleanup.

Moved helpers (all already pub(super)):

  Feedback factories:
    - build_feedback_for_tool_protocol_failure
    - build_feedback_for_deterministic_content_fallback

  Pure predicates / mappings:
    - reply_looks_like_future_work
    - answer_only_reply_is_inadequate
    - task_contract_verifier_safe_stop_mapping
    - task_contract_continue_requires_tool_recovery
    - should_apply_repo_change_partial_progress_recovery
    - should_apply_repo_change_quality_gate

  State helpers:
    - increment_artifact_completion_role_attempt
    - format_iteration_status

Adjustments:

  * actor_loop_flow.rs: added `use super::success::
    DETERMINISTIC_CONTENT_FALLBACK_TAG;` (other deps — ExitReason,
    ExecutionMode, HashMap, recovery::, format_progress_field — were
    already imported). Stripped 10 super::turn:: prefixes from
    callers that are now local.

  * turn.rs: added 4 names to the top-of-file actor_loop_flow re-
    import (build_feedback_for_deterministic_content_fallback,
    format_iteration_status, task_contract_verifier_safe_stop_mapping
    — used in production code paths). Removed now-unused
    `use std::collections::HashMap`, `use super::progress_text::
    format_progress_field`, and `use super::success::
    DETERMINISTIC_CONTENT_FALLBACK_TAG` (the consumers all moved).
    Updated 3 test mods (mod tests, mod truncate_tests, mod
    progress_tests) to import the relocated fns from
    super::super::actor_loop_flow:: explicitly. Rewrote 2
    super::build_feedback_for_tool_protocol_failure( call sites
    accordingly.

  * Orphaned doc-comment blocks above the relocated functions
    removed (clippy::empty_line_after_doc_comments fail-closed
    under -D warnings).

Metrics: turn.rs 35,322 -> 35,120 (-202); actor_loop_flow.rs
4,359 -> 4,539 (+180). Remaining gap to Issue #681 acceptance
(turn.rs <= 35,000): ~120 LOC.
